o/devicestate: move users test export to common export
