+++ overlord/devicestate/export_test.go
@@ -22,6 +22,7 @@ package devicestate
 import (
 	"context"
 	"net/http"
+	"os/user"
 	"time"
 
 	"github.com/snapcore/snapd/asserts"
@@ -30,6 +31,7 @@ import (
 	"github.com/snapcore/snapd/gadget/install"
 	"github.com/snapcore/snapd/httputil"
 	"github.com/snapcore/snapd/kernel/fde"
+	"github.com/snapcore/snapd/osutil"
 	"github.com/snapcore/snapd/overlord/snapstate"
 	"github.com/snapcore/snapd/overlord/state"
 	"github.com/snapcore/snapd/overlord/storecontext"
@@ -413,3 +415,37 @@ func MockSecbootRemoveRecoveryKeys(f func(mountPointToRecoverKeyFile map[string]
 	secbootRemoveRecoveryKeys = f
 	return restore
 }
+
+func MockOsutilAddUser(addUser func(name string, opts *osutil.AddUserOptions) error) (restore func()) {
+	oldAddUser := osutilAddUser
+	// internal.SetOsutilAddUser(addUser)
+	osutilAddUser = addUser
+	return func() {
+		osutilAddUser = oldAddUser
+		// internal.SetOsutilAddUser((func(name string, opts *osutil.AddUserOptions) error)(oldAddUser))
+	}
+}
+
+func MockOsutilDelUser(delUser func(name string, opts *osutil.DelUserOptions) error) (restore func()) {
+	oldDelUser := osutilDelUser
+	// internal.SetOsutilDelUser(delUser)
+	osutilDelUser = delUser
+	return func() {
+		// internal.SetOsutilDelUser((func(name string, opts *osutil.DelUserOptions) error)(oldDelUser))
+		osutilDelUser = oldDelUser
+	}
+}
+
+func MockUserLookup(lookup func(username string) (*user.User, error)) (restore func()) {
+	oldLookup := userLookup
+	// internal.SetUserLookup(lookup)
+	userLookup = lookup
+	return func() {
+		// internal.SetUserLookup((func(username string) (*user.User, error))(oldLookup))
+		userLookup = oldLookup
+	}
+}
+
+var (
+	GetUserDetailsFromAssertion = getUserDetailsFromAssertion
+)
